+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/repository/PrisonVisitRequestRulesRepository.kt
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.PrisonVisitRequestRules
 
 @Repository
-interface PrisonVisitRequestRulesRepository : JpaRepository<PrisonVisitRequestRules, Int> {
+interface PrisonVisitRequestRulesRepository : JpaRepository<PrisonVisitRequestRules, Long> {
   @Query(
     "SELECT r FROM PrisonVisitRequestRules r " +
       "WHERE r.prison.code = :prisonCode " +
